Add personal assistant tools implementation with task management, search, and configuration capabilities
